pyproject.toml
[project]
name = "arxiv-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.75.0",
    "openai-agents>=0.0.11",
    "python-dotenv>=1.1.0",
]

[project.scripts]
post-arxiv-summary = "main:main"
